requirements.txt
Flask
connexion[swagger-ui]
celery
SQLAlchemy
Flask-SQLAlchemy
Flask-Migrate
alembic
psycopg2-binary
requests

# Requirements needed for saving data on GCP
google-cloud-storage
